feature: No more wrapping in ScrollSyncPane and overflow:auto required
This new approach does all the heavy lifting for the user.

The user now only has to define the elements inside the `ScrollSync`
container. Those elements are automatically wrapped in an `overflow:auto`
div (ScrollSyncPane).

It also removes the usage of (discouraged by react team) `context` and
replaces it with a simple `register` prop.

It also gets rid of `ReactDOM.findNode`.